+++ Sources/TuistGraphTesting/Models/Project+TestData.swift
@@ -3,20 +3,22 @@ import TSCBasic
 @testable import TuistGraph
 
 public extension Project {
-    static func test(path: AbsolutePath = AbsolutePath("/Project"),
-                     sourceRootPath: AbsolutePath = AbsolutePath("/Project"),
-                     xcodeProjPath: AbsolutePath = AbsolutePath("/Project/Project.xcodeproj"),
-                     name: String = "Project",
-                     organizationName: String? = nil,
-                     developmentRegion: String? = nil,
-                     settings: Settings = Settings.test(),
-                     filesGroup: ProjectGroup = .group(name: "Project"),
-                     targets: [Target] = [Target.test()],
-                     packages: [Package] = [],
-                     schemes: [Scheme] = [],
-                     ideTemplateMacros: IDETemplateMacros? = nil,
-                     additionalFiles: [FileElement] = []) -> Project
-    {
+    static func test(
+        path: AbsolutePath = AbsolutePath("/Project"),
+        sourceRootPath: AbsolutePath = AbsolutePath("/Project"),
+        xcodeProjPath: AbsolutePath = AbsolutePath("/Project/Project.xcodeproj"),
+        name: String = "Project",
+        organizationName: String? = nil,
+        developmentRegion: String? = nil,
+        settings: Settings = Settings.test(),
+        filesGroup: ProjectGroup = .group(name: "Project"),
+        targets: [Target] = [Target.test()],
+        packages: [Package] = [],
+        schemes: [Scheme] = [],
+        ideTemplateMacros: IDETemplateMacros? = nil,
+        additionalFiles: [FileElement] = [],
+        resourceSynthesizers: [ResourceSynthesizer] = []
+    ) -> Project {
         Project(
             path: path,
             sourceRootPath: sourceRootPath,
@@ -30,24 +32,27 @@ public extension Project {
             packages: packages,
             schemes: schemes,
             ideTemplateMacros: ideTemplateMacros,
-            additionalFiles: additionalFiles
+            additionalFiles: additionalFiles,
+            resourceSynthesizers: resourceSynthesizers
         )
     }
 
-    static func empty(path: AbsolutePath = AbsolutePath("/test/"),
-                      sourceRootPath: AbsolutePath = AbsolutePath("/test/"),
-                      xcodeProjPath: AbsolutePath = AbsolutePath("/test/text.xcodeproj"),
-                      name: String = "Project",
-                      organizationName: String? = nil,
-                      developmentRegion: String? = nil,
-                      settings: Settings = .default,
-                      filesGroup: ProjectGroup = .group(name: "Project"),
-                      targets: [Target] = [],
-                      packages: [Package] = [],
-                      schemes: [Scheme] = [],
-                      ideTemplateMacros: IDETemplateMacros? = nil,
-                      additionalFiles: [FileElement] = []) -> Project
-    {
+    static func empty(
+        path: AbsolutePath = AbsolutePath("/test/"),
+        sourceRootPath: AbsolutePath = AbsolutePath("/test/"),
+        xcodeProjPath: AbsolutePath = AbsolutePath("/test/text.xcodeproj"),
+        name: String = "Project",
+        organizationName: String? = nil,
+        developmentRegion: String? = nil,
+        settings: Settings = .default,
+        filesGroup: ProjectGroup = .group(name: "Project"),
+        targets: [Target] = [],
+        packages: [Package] = [],
+        schemes: [Scheme] = [],
+        ideTemplateMacros: IDETemplateMacros? = nil,
+        additionalFiles: [FileElement] = [],
+        resourceSynthesizers: [ResourceSynthesizer] = []
+    ) -> Project {
         Project(
             path: path,
             sourceRootPath: sourceRootPath,
@@ -61,7 +66,8 @@ public extension Project {
             packages: packages,
             schemes: schemes,
             ideTemplateMacros: ideTemplateMacros,
-            additionalFiles: additionalFiles
+            additionalFiles: additionalFiles,
+            resourceSynthesizers: resourceSynthesizers
         )
     }
 }