+++ Sources/TuistLoader/Models+ManifestMappers/SwiftPackageManagerDependencies+ManifestMapper.swift
@@ -1,0 +1,18 @@
+import Foundation
+import ProjectDescription
+import TSCBasic
+import TuistCore
+import TuistGraph
+import TuistSupport
+
+extension TuistGraph.SwiftPackageManagerDependencies {
+    /// Creates `TuistGraph.SwiftPackageManagerDependencies` instance from `ProjectDescription.SwiftPackageManagerDependencies` instance.
+    static func from(
+        manifest: ProjectDescription.SwiftPackageManagerDependencies,
+        generatorPaths: GeneratorPaths
+    ) throws -> Self {
+        let packages = try manifest.packages.map { try TuistGraph.Package.from(manifest: $0, generatorPaths: generatorPaths) }
+
+        return .init(packages)
+    }
+}