+++ Sources/TuistSupport/Constants.swift
@@ -6,7 +6,7 @@ public enum Constants {
     public static let binName = "tuist"
     public static let gitRepositoryURL = "https://github.com/tuist/tuist.git"
     public static let communityURL = "https://community.tuist.io"
-    public static let version = "1.39.1"
+    public static let version = "1.40.0"
     public static let bundleName: String = "tuist.zip"
     public static let trueValues: [String] = ["1", "true", "TRUE", "yes", "YES"]
     public static let tuistDirectoryName: String = "Tuist"
@@ -18,6 +18,7 @@ public enum Constants {
     public static let masterKey = "master.key"
     public static let encryptedExtension = "encrypted"
     public static let templatesDirectoryName: String = "Templates"
+    public static let resourceSynthesizersDirectoryName: String = "ResourceSynthesizers"
     public static let vendorDirectoryName: String = "vendor"
     public static let twitterHandle: String = "tuistio"
     public static let joinSlackURL: String = "https://slack.tuist.io/"
@@ -33,7 +34,9 @@ public enum Constants {
         public static let name = "Dependencies"
         public static let lockfilesDirectoryName = "Lockfiles"
         public static let cartfileResolvedName = "Cartfile.resolved"
+        public static let packageResolvedName = "Package.resolved"
         public static let carthageDirectoryName = "Carthage"
+        public static let swiftPackageManagerDirectoryName = "SwiftPackageManager"
     }
 
     public enum DerivedDirectory {